+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes on `main` will be listed here.
 
 ### Features
 
+-   [[#14](https://github.com/diatche/react-native-form-model/pull/14)] Added `autoFocus` to `InputFieldModel`.
 -   [[#13](https://github.com/diatche/react-native-form-model/pull/13)] Input labels are now accessible.
 -   [[#13](https://github.com/diatche/react-native-form-model/pull/13)] Select next form field on editing submit.
     -   It's possible to skip fields when editing is submitted on a field with the `skipNextFocus` property.